pyproject.toml
[project]
name = "merg"
version = "0.0.2"
description = "Deep merging for Python with strict JSON/YAML-shaped type validation."
authors = [{name = "Freedom Fury", email = "freeedomfury@gmail.com"}]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/freedomfury/merg"
Repository = "https://github.com/freedomfury/merg"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["merg"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
